requirements.txt
python-pptx==0.6.21
Pillow==10.0.1
pytesseract==0.3.10
requests==2.31.0
tqdm==4.66.1
PyYAML==6.0.1
python-dotenv==1.0.0 